Commit of second home task - application with 2 modules.
My implementation simulates the intercom system of apartment house. So,
it has one calling unit and many subscriber units. Due to one-to-many
model, current implementation has major issue related to creating of
many instances of same SubscriberUnit class (I couldn't make the methods
and properties private in SubscriberUnit). This need to be discussed
with course leader.